Changed local SDK to Python3.11
I tried to make this change on Digital Ocean as well, but the function had trouble loading
the imported modules even though they are all clearly installing during the remote build.
I don't have time to troubleshoot this now so I moved the deployment back to python 3.9
but I intend to get this resolved in the future so we can use Python 3.11 as the runtime.