Completed update to new workflow
As described by the flowchart from Whimsical, the opt-in process is now more clear and
the subsequent messages are only sent once the opt-in is complete and recorded. I put
some initial thought into how we will want to handle messages that are delayed by
greater than 30 minutes in case we want to handle that separately. There are still a
number of places in the code where improvements should be made, but this is a good
stopping point for now, and likely a place where we can do some testing with the rest
of the EDC team.